rename from cwc_helper to cwc_ext
