+++ lib/makefile
@@ -4,7 +4,7 @@ include config.mk
 CFLAGS := -O3 -ffast-math -Wall -I"." $(CFLAGS)
 NVFLAGS := -O3 $(NVFLAGS)
 
-NNC_OBJS := nnc/*.o nnc/cpu/*.o
+NNC_OBJS := nnc/ccv_nnc.o nnc/cpu/ccv_nnc_ref.o
 
 .PHONY: all lib clean
 
@@ -27,8 +27,8 @@ ccv_io.o: ccv_io.c ccv.h ccv_internal.h io/*.c
 ../samples/image-net-2012-vgg-d.sqlite3:
 	../samples/download-vgg-d-model.sh
 
-%.o: %.c ccv.h ccv_internal.h cuda/cwc.h cuda/cwc_helper.h nnc/ccv_nnc_tfb.h
+%.o: %.c ccv.h ccv_internal.h cuda/cwc.h cuda/cwc_ext.h nnc/ccv_nnc_tfb.h
 	$(CC) $< -o $@ -c $(CFLAGS)
 
-%.o: %.cu ccv.h cuda/cwc.h cuda/cwc_internal.h cuda/cwc_helper.h nnc/ccv_nnc_tfb.h
+%.o: %.cu ccv.h cuda/cwc.h cuda/cwc_internal.h cuda/cwc_ext.h nnc/ccv_nnc_tfb.h
 	$(NVCC) $< -o $@ -c $(NVFLAGS)
